Plot data with archetypes in 2D and 3D

plot_arc() plot data with polytope representing the Pareto front, where vertices are archetypes (dots connected with lines). When archetype data is "b_pch_fit" all archetype locations from each subsample are shown with lines connecting the average location (type "average"); or lines connecting archetypes in each of the experiments (colored differently, type "all").

plot_arc(arch_data = NULL, data, which_dimensions = as.integer(1:2),
  type = c("average", "all")[1], average_func = mean,
  geom = list(ggplot2::geom_point, ggplot2::geom_bin2d)[[1]],
  colors = c("#1F77B4", "#D62728", "#2CA02C", "#17BED0", "#006400",
  "#FF7E0F"), arch_size = NULL, line_size = NULL, data_size = 2,
  arch_alpha = 0.4, data_lab = "data", arc_lab = "archetypes",
  legend_name = "data", text_size = NULL, nudge = c(0.05, 0.1))

Arguments

data
matrix of data in which archetypes/polytope were found, dim(variables/dimentions, examples)

which_dimensions
indices or character vector specifying dimension names. When which_dimensions exceeds the number of dimensions in arc_data these archetypes will be omitted. This can happen when fitting simplexes: lines and triangles are only 2D, so will be omitted from 3D plots.

type
used when arc_data is "b_pch_fit", one of "average", "all"

average_func
used when arc_data is "b_pch_fit", function telling how to find average position of vertices

geom
plotting function to plot data in 2D, useful options are ggplot2::geom_point (scatterplot) and ggplot2::geom_bin2d (density)

colors
character vector giving color palette for different archetype fits and the data (both 3D and 2D plot)

arch_size
size of archetype points

line_size
width of lines connecting archetypes

data_size
size of data points in plotly. Values for ggplot are 1/2 of data_size.

arch_alpha
opacity of archetype points

data_lab
vector, 1L or length of data, label data points (examples) with a qualitative or quantitative label

arc_lab
vector, 1L or nrow(arc_data$XC)/noc, label vertices/archetypes (points) with a qualitative or quantitative label

legend_name
name to display on legend, e.g. gene name in data_lab

text_size
archetype label text size

arc_data
objects of class "pch_fit", "b_pch_fit", "k_pch_fit" storing the position of archetypes, and other data from fit_pch() run. arc_data$XC is matrix of dim(dimensions, archetypes) or list where each element is XC matrix from an independent run of the polytope fitting algorithm. Set to NULL if you want to show data alone.

Value

plot_arc() ggplot2 (2D) or plotly (3D) plot

See also

Examples

library(ParetoTI)
library(ggplot2)
# Random data that fits into the triangle (2D)
set.seed(4355)
archetypes = generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),
                          mean = 0, sd = 1, N_dim = 2)

#> Error in generate_arc(arc_coord = list(c(5, 0), c(-10, 15), c(-30, -20)),     mean = 0, sd = 1, N_dim = 2): unused argument (N_dim = 2)

data = generate_data(archetypes$XC, N_examples = 1e4, jiiter = 0.04, size = 0.9)

#> Error in ncol(archetypes): object 'archetypes' not found

plot_arc(arch_data = archetypes, data = data,
    which_dimensions = 1:2) +
    theme_bw()

#> Error in plot_arc(arch_data = archetypes, data = data, which_dimensions = 1:2): object 'archetypes' not found

# Plot data as 2D density rather than points
plot_arc(arch_data = archetypes, data = data,
    which_dimensions = 1:2, geom = ggplot2::geom_bin2d)

#> Error in plot_arc(arch_data = archetypes, data = data, which_dimensions = 1:2,     geom = ggplot2::geom_bin2d): object 'archetypes' not found

# Random data that fits into the triangle (3D)
set.seed(4355)
archetypes = generate_arc(arc_coord = list(c(5, 0, 4), c(-10, 15, 0), c(-30, -20, -5)),
                          mean = 0, sd = 1, N_dim = 3)

#> Error in generate_arc(arc_coord = list(c(5, 0, 4), c(-10, 15, 0), c(-30,     -20, -5)), mean = 0, sd = 1, N_dim = 3): unused argument (N_dim = 3)

data = generate_data(archetypes$XC, N_examples = 1e4, jiiter = 0.04, size = 0.9)

#> Error in ncol(archetypes): object 'archetypes' not found

plot_arc(arch_data = archetypes, data = data,
    which_dimensions = 1:3) +
    theme_bw()

#> Error in plot_arc(arch_data = archetypes, data = data, which_dimensions = 1:3): object 'archetypes' not found