Plot gam model fit for multiple genes and archetypes (Generalised Additive Model, mgcv)

plot_gam() plots gam model fit for multiple features (genes) and archetypes

plot_gam(gam_deriv = NULL, gam_fit, data, feature = NULL,
  groupCovs = NULL, title_size = 10)

Arguments

gam_deriv
object with multiple gam fits (many arhetypes, many features). Either gam_deriv or gam_fit must be provided.

gam_fit
a fitted gam object as produced by gam().

data
data.table dim(examples, dimensions) that includes distance of each example to archetype in columns given by arc_col and feature values given by feature. Must be provided.

feature
character vector (1L), column than containg feature values

groupCovs
argument for plotGAM

Value

plot_gam() list (S3 object, plot_gam) containing data.table-s for each predictor in gam model

Details

Plotting of multiple gam fits is inspired by https://stackoverflow.com/questions/49471300/gam-plots-with-ggplot